large amount of bug fixing went on here... so first when throtte is set to zero the integral acummilation is set to zero as the pid is not active during this time so it just accumilated and ramped up values as the pid was trying to compensate but the motors werent acting. the next problem was the gyro, in our tests today we were never getting gyro data! idk how this happened or who commented it out but it is now back in so we have the correct values. we reoriented the gyro to be opposite to the the inteded facing angle on the drone so we were going to invert the angles that we recived from the gyro so we reverted back to the inteded values. finally FLOATS! in the pwm code I was being too eager to convert the duty values to an int so when multiplying by a float we were just getting stupid values. this got changed to do the convesion at last minuite and that worked beautifully

+++ ControlIntegration/ESCPID/AccelGyro/PWM.h
@@ -3,7 +3,7 @@
 #include "Definitions.h"
 #include "Arduino.h"
 void init_pwm(void);
-void pwm_duty(uint8_t motor,uint16_t duty);
+void pwm_duty(uint8_t motor,float duty);
 void setMotors (float throttle, float yaw, float pitch, float roll);
 
 #endif
